Changed decimal to double, changed your calculations to precision float, using meters now

+++ src/spatial/attitude-utils.hpp
@@ -7,7 +7,7 @@ namespace found {
 
 /// Alias for floating point numbers. Used for controlling
 /// floating-point type memory usage throughout the program
-typedef float decimal;
+typedef double decimal;
 
 // At first, I wanted to have two separate Attitude classes, one storing Euler angles and converting
 // to Quaterinon, and another storing as Quaternion and converting to Euler. But abstract classes
